+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/booking/BookingStatus.java
@@ -1,0 +1,13 @@
+package com.pgms.coredomain.domain.booking;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum BookingStatus {
+
+	WAITING_FOR_DEPOSIT("입금대기"),
+	PAYMENT_COMPLETED("결제완료"),
+	CANCELLED("취소");
+
+	private final String description;
+}
