+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/booking/PaymentMethod.java
@@ -1,13 +1,12 @@
-package com.pgms.coredomain.domain.order;
+package com.pgms.coredomain.domain.booking;
 
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
 public enum PaymentMethod {
 
     CARD("카드"),
     VIRTUAL_ACCOUNT("가상 계좌");
 
     private final String description;
-
-    PaymentMethod(String description) {
-        this.description = description;
-    }
 }
