+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/booking/ReceiptType.java
@@ -1,0 +1,11 @@
+package com.pgms.coredomain.domain.booking;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum ReceiptType {
+	PICK_UP("현장수령"),
+	DELIVERY("배송");
+
+	private final String description;
+}
